+++ src/tko/game/game.py
@@ -78,7 +78,6 @@ class Game:
 
     def load_sources(self, sources: list[RepSource], language: str):
         for source in sources:
-            # print("debug loading", source.alias, source.cache_path)
             filename = source.target_path
             filters = source.filters
             content = ""